Rename reforms to fit documentation
