+++ python/reforme_parametrique.py
